#1496 renamed smv.io to smv.iomod to avoid conflect with top level io package
