--- src/main/python/smv/io/__init__.py
@@ -1,2 +1,0 @@
-from smv.io.inputs import *
-from smv.io.outputs import *
