+++ src/main/python/smv/iomod/__init__.py
@@ -1,0 +1,2 @@
+from smv.iomod.inputs import *
+from smv.iomod.outputs import *
