+++ src/main/python/smv/iomod/outputs.py
@@ -11,21 +11,39 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from smv.io.base import SmvInput, AsTable
+from smv.iomod.base import SmvSparkDfOutput, AsTable
 
 
-class SmvJdbcInputTable(SmvInput, AsTable):
+class SmvJdbcOutputTable(SmvSparkDfOutput, AsTable):
     """
         User need to implement 
 
+            - requiresDS
             - connectionName
             - tableName
+            - writeMode: optional, default "errorifexists"
     """
 
+    def writeMode(self):
+        """
+            Write mode for Spark DataFrameWriter.
+            Valid values:
+
+                - "append"
+                - "overwrite"
+                - "ignore"
+                - "error" or "errorifexists" (default)
+        """
+        return "errorifexists"
+
+
     def doRun(self, known):
+        data = self.get_spark_df(known)
         conn = self.get_connection()
-        builder = self.smvApp.sqlContext.read\
-            .format('jdbc')\
+
+        builder = data.write\
+            .format("jdbc") \
+            .mode(self.writeMode()) \
             .option('url', conn.url)
 
         if (conn.driver is not None):
@@ -35,10 +53,15 @@ class SmvJdbcInputTable(SmvInput, AsTable):
         if (conn.password is not None):
             builder = builder.option('password', conn.password)
 
-        return builder\
-            .option('dbtable', self.tableName())\
-            .load()
+        builder \
+            .option("dbtable", self.tableName()) \
+            .save()
+
+        # return data back for meta calculation
+        # TODO: need to review whether we should even calculate meta for output
+        return data
+
 
 __all__ = [
-    'SmvJdbcInputTable',
+    'SmvJdbcOutputTable',
 ]
